refactor: update PipelineMiddleware
