+++ src/WeihanLi.Common/Helpers/Pipelines/PipelineBuilder.cs
@@ -45,21 +45,3 @@ internal sealed class PipelineBuilder<TContext> : IPipelineBuilder<TContext>
     public IPipelineBuilder<TContext> New() => new PipelineBuilder<TContext>(_completeFunc);
 }
 
-public interface IPipelineMiddleware<TContext>
-{
-    void Invoke(TContext context, Action<TContext> next);
-}
-
-public interface IAsyncPipelineMiddleware<TContext>
-{
-    Task InvokeAsync(TContext context, Func<TContext, Task> next);
-}
-
-#if ValueTaskSupport
-
-public interface IValueAsyncPipelineMiddleware<TContext>
-{
-    ValueTask InvokeAsync(TContext context, Func<TContext, ValueTask> next);
-}
-
-#endif
